requirements.txt
# List of dependencies for the project
# Use `pip install -r requirements.txt` to install these dependencies

# Add your dependencies below, for example:
# flask==2.0.3  # Web framework for building applications
# requests==2.28.1  # HTTP library for making API calls
# opentelemetry-api==1.5.0  # OpenTelemetry API for distributed tracing
# opentelemetry-sdk==1.5.0  # OpenTelemetry SDK for distributed tracing

fastapi
uvicorn
requests

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests

prometheus-client